fix: move HTML template file to dedicated statics folder (#270)
